fix&move: 로컬 경로 이미지 버그 수정 및 파일 이동
